ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
transformers==4.35.2
torch>=2.2.0
scikit-learn==1.3.2
spacy==3.7.2
pandas==2.1.3
numpy==1.25.2
openai==1.3.5
python-dotenv==1.0.0
hypothesis==6.88.1
pytest==7.4.3
reportlab==4.0.7
PyPDF2==3.0.1